Add files via upload
Hypertext Markup Language is the abbreviation for HTML. It is used for developing the front-end component of a web page. A web page is HTML with Hypertext and Markup. Links between web pages are defined by Hypertext. HTML markup language is used to define the structure of a web page's text documentation.

A CSS file, also known as Cascading Style Sheets, is a specially designed language intended to simplify the process of creating attractive web pages. It allows you to add styles to your page. In addition, CSS enables you to do this.

The JavaScript library React plays a declarative, efficient, and flexible role in building user interfaces. In an application,ReactJS is responsible solely for the view layer, as it is a component-based library.

jQuery will also be used for the UI design